# Add `index_sort` implementation