--- a/pyg_lib/csrc/ops/index_sort.h
+++ b/pyg_lib/csrc/ops/index_sort.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <ATen/ATen.h>
+#include <cstdint>
+#include <tuple>
+#include "pyg_lib/csrc/macros.h"
+
+namespace pyg {
+namespace ops {
+
+PYG_API std::tuple<at::Tensor, at::Tensor> index_sort(
+    const at::Tensor& input,
+    const at::optional<int64_t> max);
+
+}  // namespace ops
+}  // namespace pyg