--- a/test/ops/test_index_sort.py
+++ b/test/ops/test_index_sort.py
@@ -22,7 +22,5 @@ def test_index_sort(device):
 @pytest.mark.parametrize('device', DEVICES)
 def test_index_sort_negative(device):
     input = torch.randint(low=0, high=1024, size=(16, 32), device=device)
-    # this should fail, as we do not support ndim > 1
-    # check in pyg_lib/csrc/ops/index_sort.cpp is not performed
-    # TODO: fix this
-    sorted_input, indices = pyg_lib.ops.index_sort(input)
+    with pytest.raises(RuntimeError):
+        sorted_input, indices = pyg_lib.ops.index_sort(input)